apps/reddit-meaningofwords/requirements.txt
openai==1.76.0
praw==7.8.1
pydantic==2.11.3
python-dotenv==1.1.0
